refactor: separate 2 func from server.ts
